모든 부분 연결 완료 그에 따른 코드 수정 및 질문/의견 제시

+++ src/main/kotlin/com/sparta/techTree/post/service/PostServiceImpl.kt
@@ -16,6 +16,7 @@ import com.sparta.techTree.user.model.UserEntity
 import com.sparta.techTree.user.repository.UserRepository
 import jakarta.persistence.EntityNotFoundException
 import org.springframework.security.core.context.SecurityContextHolder
+import java.nio.file.AccessDeniedException
 
 
 @Service
@@ -42,10 +43,7 @@ class PostServiceImpl(private val postRepository: PostRepository, private val li
 
 
     override fun createPost(request: CreatePostRequest, userId: Long): PostResponse {
-        val user = userRepository.findById(userId)
-            .orElseThrow {
-                EntityNotFoundException("User with ID $userId not found")
-            }
+        val user: UserEntity = userRepository.findByIdOrNull(userId) ?: throw ModelNotFoundException("User",userId)
         val createdPost = postRepository.save(
             Post(
                 title = request.title,
@@ -60,6 +58,11 @@ class PostServiceImpl(private val postRepository: PostRepository, private val li
     @Transactional
     override fun updatePost(postId: Long, userId: Long, request: UpdatePostRequest,): PostResponse {
         val post = postRepository.findByIdOrNull(postId) ?: throw ModelNotFoundException("Post", postId)
+        //이 if문이 없으면 userid 1을 가진사람이 userid 2인 사람의 글도 수정이 가능함
+        //저는 이랬는데, 혹시 이거 없이도 잘 되신분 있으시면 말씀해주세요!
+        if (post.user.id != userId) {
+            throw AccessDeniedException("User with ID $userId does not have permission to update post with ID $postId")
+        }
         val countLikes = likeRepository.countByPostId(postId)
         post.title = request.title ?: post.title
         post.content = request.content ?: post.content
@@ -70,6 +73,11 @@ class PostServiceImpl(private val postRepository: PostRepository, private val li
     @Transactional
     override fun deletePost(postId: Long,userId: Long) {
         val post = postRepository.findByIdOrNull(postId) ?: throw ModelNotFoundException("Post", postId)
+        //이 if문이 없으면 userid 1을 가진사람이 userid 2인 사람의 글도 삭제가 가능함
+        //저는 이랬는데, 혹시 이거 없이도 잘 되신분 있으시면 말씀해주세요!
+        if (post.user.id != userId) {
+            throw AccessDeniedException("User with ID $userId does not have permission to update post with ID $postId")
+        }
         postRepository.delete(post)
     }
 }
